requirements.txt
Django==1.9.6
djangorestframework==3.3.3
feedparser==5.2.1
iso8601==0.1.11
